shoe_store/requirements.txt
Django==4.2.7
Pillow==10.1.0
pandas==2.1.3
openpyxl==3.1.2
django-cleanup==8.0.0
psycopg2-binary